+++ librecad/src/ui/forms/qg_circletan2options.cpp
@@ -66,7 +66,7 @@ void QG_CircleTan2Options::setAction(RS_ActionInterface* a, bool update) {
 
         QString sr;
         if (update) {
-            sr = QString("%1").arg(action->getRadius());
+			sr = QString("%1").arg((double)action->getRadius());
         } else {
             RS_SETTINGS->beginGroup("/Draw");
             sr = RS_SETTINGS->readEntry("/CircleTan2Radius", "1.0");
