+++ librecad/src/ui/forms/qg_moverotateoptions.cpp
@@ -53,7 +53,7 @@ void QG_MoveRotateOptions::setAction(RS_ActionInterface* a, bool update) {
 
         QString sa;
         if (update) {
-            sa = QString("%1").arg(RS_Math::rad2deg(action->getAngle()));
+			sa = QString("%1").arg((double)RS_Math::rad2deg(action->getAngle()));
         } else {
             RS_SETTINGS->beginGroup("/Modify");
             sa = RS_SETTINGS->readEntry("/MoveRotate", "30");
